refactor(finalizer): Remove opStack code from index
This cleans up index.ts and removes a lot of exports from
utils/opStack.ts.

Have only done Optimism so far to keep the diff simple, but will repeat
separately for Polygon and Arbtirum.
